Abtract Rotating Handler wtih implementation RotatingFileHandler
